[MOVE] PLUBError code
